requirements.txt
python-dotenv
pandas
rich
google-generativeai
requests